fix(velociraptor): satisfy lint + file-size CI gates from the title fix
The previous commit's non-null assertion tripped
@typescript-eslint/no-unnecessary-type-assertion (TS narrows `flat` from
the `flat && isObject(...)` guard, so the assertion was redundant) and
pushed velociraptorImport.ts 39 lines past its file-size ledger cap.

Split the three small pure title-formatting helpers (withHostSuffix,
titleSafe, isMangledUtf16) out into velociraptorTitle.ts rather than
growing the already-oversized file further, and fixed the narrowing so
no assertion is needed at all.
